Use tty system command on Unix platforms except for Linux (#895)
On Linux, it uses symbolic link of proc filesystem as previous, which is
not portable on other Unix platforms.
On Windows, it just returns empty string.
