+++ pkg/backend/crypto/gpg/cli/utils.go
@@ -7,10 +7,6 @@ import (
 	"time"
 )
 
-var (
-	fd0 = "/proc/self/fd/0"
-)
-
 // parseTS parses the passed string as an Epoch int and returns
 // the time struct or the zero time struct
 func parseTS(str string) time.Time {
@@ -46,15 +42,6 @@ func splitPacket(in string) map[string]string {
 		m[p[i]] = strings.Trim(p[i+1], ",")
 	}
 	return m
-}
-
-// see https://www.gnupg.org/documentation/manuals/gnupg/Invoking-GPG_002dAGENT.html
-func tty() string {
-	dest, err := os.Readlink(fd0)
-	if err != nil {
-		return ""
-	}
-	return dest
 }
 
 // GPGOpts parses extra GPG options from the environment
