+++ pkg/backend/crypto/gpg/cli/utils_linux_test.go
@@ -1,0 +1,14 @@
+// +build linux
+
+package cli
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestTTY(t *testing.T) {
+	fd0 = "/tmp/foobar"
+	assert.Equal(t, "", tty())
+}
