+++ pkg/backend/crypto/gpg/cli/utils_windows.go
@@ -1,0 +1,7 @@
+// +build windows
+
+package cli
+
+func tty() string {
+	return ""
+}
